Move example parameter set to own files
