+++ ewatercycle/parameter_sets/_lisflood.py
@@ -1,5 +1,6 @@
-from typing import Optional
+from typing import Optional, Iterable
 
+from ._example import ExampleParameterSet
 from ._default import ParameterSet, make_absolute
 
 
@@ -25,3 +26,20 @@ class LisfloodParameterSet(ParameterSet):
     @property
     def is_available(self):
         return super().is_available and self.MaskMap.exists()
+
+
+def example_parameter_sets() -> Iterable[ExampleParameterSet]:
+    return [
+        ExampleParameterSet(
+            # Relative to CFG['parameterset_dir']
+            directory="lisflood_fraser",
+            name="lisflood_fraser",
+            # Relative to CFG['parameterset_dir']
+            config="lisflood_fraser/settings_lat_lon-Run.xml",
+            datafiles_url="https://github.com/ec-jrc/lisflood-usecases/trunk/LF_lat_lon_UseCase",
+            # Raw url to config file
+            config_url="https://github.com/ec-jrc/lisflood-usecases/raw/master/LF_lat_lon_UseCase/settings_lat_lon-Run.xml",
+            doi="N/A",
+            target_model="lisflood",
+        )
+    ]
